Fix sending lists of data inside the argument list
Since we can't do varargs in Erlang, we use lists:
  send(Cxn, Key, [Value1, Value2, Value3])
But sometimes we want to be lazy and just do:
  send(Cxn, Key, "set a value")

This patch allows both cases to co-exist.  Instead of (as
previously) crashing, we convert the inner values to
binaries then return the converted list.  At the end, the deep
list of binaries is flattened to perform redis protocol
calculations and construction.